+++ lib/constants/config.dart
@@ -1,1 +1,4 @@
-const String urlRepo = 'http://103.82.241.80:3000';
+// DEBUG
+const String urlRepo = 'http://103.82.241.80:3001';
+// PRODUCTION
+// const String urlRepo = 'http://103.82.241.80:3000';