+++ lib/helpers/object_to_db_helper.dart
@@ -36,13 +36,12 @@ EmployeeDB getEmployeeDBFromEmployee(Employee employee) {
 ReportRegTorqueDB getReportTorqueDBFromMasterReportTorque(
         MasterReportRegTorque torque, index) =>
     ReportRegTorqueDB(
-      towerSegment: torque.towerSegment,
-      elevasi: torque.elevasi,
-      boltSize: torque.boltSize,
-      minimumTorque: torque.minimumTorque,
-      qtyBolt: torque.qtyBolt,
-      orderIndex: index
-    );
+        towerSegment: torque.towerSegment,
+        elevasi: torque.elevasi,
+        boltSize: torque.boltSize,
+        minimumTorque: torque.minimumTorque,
+        qtyBolt: torque.qtyBolt,
+        orderIndex: index);
 
 ReportRegVerticalityDB getReportRegVerticalityDBFromReportRegVerticality(
         ReportRegVerticality report) =>