+++ lib/screen/components/task_card.dart
@@ -4,6 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 import 'package:monitor_app/constants/images.dart';
+import 'package:monitor_app/helpers/format_helper.dart';
 import 'package:monitor_app/model/task.dart';
 import 'package:monitor_app/screen/components/icon_task_status.dart';
 
@@ -37,14 +38,44 @@ class TaskCard extends ConsumerWidget {
                     height: 14,
                   ),
                   const SizedBox(width: 10),
-                  Text(task.createdDate)
+                  Text(
+                    formatTanggalIndonesia(task.created_at, 2),
+                    style: const TextStyle(fontWeight: FontWeight.bold),
+                  )
                 ],
               ),
               IconTaskStatus(taskStatus: task.status),
             ],
           ),
           const SizedBox(height: 10),
-          Text('${task.site.id} - ${task.site.name}'),
+          Text(
+            '${task.site.id} - ${task.site.name}',
+            style: const TextStyle(fontSize: 16),
+          ),
+          const SizedBox(height: 10),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Container(
+                padding: const EdgeInsets.symmetric(vertical: 2, horizontal: 8),
+                decoration: BoxDecoration(
+                    color: Colors.blue[800],
+                    borderRadius: BorderRadius.circular(5)),
+                child: task.dueDate != null
+                    ? Text(
+                        'Due Date: ${formatTanggalIndonesia(task.dueDate!, 3)}',
+                        style: const TextStyle(color: Colors.white),
+                      )
+                    : const Text('-'),
+              ),
+              task.dueDate != null
+                  ? Text(
+                      '${selisihTanggal(task.dueDate!)}d left',
+                      style: const TextStyle(color: Colors.red, fontWeight: FontWeight.bold),
+                    )
+                  : const Text('-'),
+            ],
+          ),
           // const SizedBox(height: 20),
           // Row(
           //   mainAxisAlignment: MainAxisAlignment.spaceBetween,
@@ -67,19 +98,19 @@ class TaskCard extends ConsumerWidget {
     );
   }
 
-  SvgPicture _buildTodoIcon(String todoType) {
-    if (todoType.toLowerCase() == 'regular') {
-      return SvgPicture.asset(
-        icTaskRegular,
-        width: 32,
-        height: 32,
-      );
-    } else {
-      return SvgPicture.asset(
-        icTaskPreventive,
-        width: 32,
-        height: 32,
-      );
-    }
-  }
+  // SvgPicture _buildTodoIcon(String todoType) {
+  //   if (todoType.toLowerCase() == 'regular') {
+  //     return SvgPicture.asset(
+  //       icTaskRegular,
+  //       width: 32,
+  //       height: 32,
+  //     );
+  //   } else {
+  //     return SvgPicture.asset(
+  //       icTaskPreventive,
+  //       width: 32,
+  //       height: 32,
+  //     );
+  //   }
+  // }
 }